backend/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2
django-colorfield==0.10.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.1.0
pluggy==0.13.1
psycopg2-binary==2.9.9
py==1.11.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.26.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
toml==0.10.2
urllib3==1.26.16
